Split MiqAeMethodService::MiqAeServiceObject into it's own file
