+++ lib/miq_automation_engine/engine/miq_ae_method_service.rb
@@ -1,4 +1,5 @@
 require_relative 'miq_ae_method_service/deprecation'
 require_relative 'miq_ae_method_service/miq_ae_service_front'
+require_relative 'miq_ae_method_service/miq_ae_service_object'
 module MiqAeMethodService
 end
